first test passed

+++ src/main/scala/rca/auditor/LinkageCharacteristic.scala
@@ -5,12 +5,12 @@ import scala.beans.BeanProperty
 class LinkageCharacteristic (var _name: String) extends Serializable {
 
   @BeanProperty var name: String = _name
-  @BeanProperty var sample: Int = _
+  @BeanProperty var sample: Long = _
   @BeanProperty var methodology: String = _
-  @BeanProperty var score: Float = _
-  @BeanProperty var min: Float = _
-  @BeanProperty var max: Float = _
-  @BeanProperty var avg: Float = _
-  @BeanProperty var stddev: Float = _
-  @BeanProperty var mean: Float = _
+  @BeanProperty var score: Double = _
+  @BeanProperty var min: Double = _
+  @BeanProperty var max: Double = _
+  @BeanProperty var avg: Double = _
+  @BeanProperty var stddev: Double = _
+  @BeanProperty var mean: Double = _
 }
